chore: Remove unused code and rename the AboutCard component
