+++ src/components/AboutCard/styles.module.css
@@ -11,7 +11,6 @@
 }
 
 .image {
-  /* max-width: 100px; */
   border-radius: 50%;
 }
